Tasks are now kicked-off synchronously
Previously, Task functions are scheduled to run on the next micor-task
"tick":

```js
let called = false;

new Task(async run => {
  called = true;
  // ...
});

console.log(called); // => false;
console.log(task.isScheduled); // => true
console.log(task.isRunning); // => false
```

This design was based on a misunderstanding of the ember-concurrency
API. Since e-c 0.8, tasks are executed synchronously. This matches the
semantics of Promises and async functions. This is also important
because some work are timing sensitive and needs to be started as soon
as possible or within a certain window. For example, `window.open()`
generally needs to be executed within the same "tick" as the click
event callback to avoid being blocked by browsers/pop-up blockers.

This commit changed to semantics align with ember-concurrency:

```js
let called = false;

new Task(async run => {
  called = true;
  // ...
});

console.log(called); // => true;
console.log(task.isRunning); // => true
```

This also allowed us to remove the `isScheduled` and `isPending` states
and simplify some of the internals.

Note that the even if the function synchronously return, the `Task`
itself still won't be resolved until the next micro-task "tick". This
means that if you create a `Task` and immediately call `cancel()` on it
the task will be canceled (and rejected) as expected (but some of the
work might already have happened).

(Thanks @machty for the pointers.)